Completed read-only tests and per class instance creation
In order to allow the tests to run in parallel, the tests need to all be readonly, and each test file need to set up its own reference to the ADLX library.

This will get a lot more complicated when we add future tests to actually change the configuration, as those changes will need to be made serially. I will need to investigate how the xunit test suite can run a whole lot of parallel tests, then run the tests that modify things in sequence afterwards.